src/requirements.txt
datasets==2.9.0
Flask==2.2.2
Flask_MySQLdb==1.0.1
matplotlib==3.6.3
numpy==1.24.1
openai==0.26.4
pandas==1.5.3
PyJWT==2.6.0
pytest==7.2.1
bcrypt==4.0.1
python-dotenv==0.21.1
#python_bcrypt==0.3.2
requests==2.28.2
scikit_learn==1.2.1
snscrape==0.5.0.20230113
torch==1.13.1
transformers==4.26.0
